Move PTC class to drivers module
